backend2/requirementsNO.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
asn1crypto==1.4.0
asttokens==2.0.5
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
autopep8==1.6.0
base58==2.1.1
bitarray==1.2.2
black==21.11b1
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
coincurve==17.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cytoolz==0.11.2
dataclassy==0.11.1
defusedxml==0.7.1
Django==3.2.7
django-cors-headers==3.8.0
django-pdfkit==0.3.1
django-templated-mail==1.1.1
django-wkhtmltopdf==3.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
eip712==0.1.0
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.2
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
ethereum==2.3.2
execnet==1.9.0
frozenlist==1.2.0
future==0.18.2
hexbytes==0.2.2
hypothesis==6.27.3
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.0.3
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
oauthlib==3.2.0
packaging==21.3
parsimonious==0.8.1
pathspec==0.9.0
pbkdf2==1.3
pdfkit==0.6.0
Pillow==9.0.0
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.23
protobuf==3.19.1
psutil==5.8.0
psycopg2==2.9.3
py==1.11.0
py-ecc==6.0.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.12.0
pyethash==0.1.27
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
pysha3==1.0.2
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.19.2
python3-openid==3.2.0
pythx==1.6.1
pytz==2021.3
PyYAML==5.4.1
regex==2021.11.10
reportlab==3.6.6
repoze.lru==0.7
requests==2.26.0
requests-oauthlib==1.3.1
rlp==2.0.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scrypt==0.8.19
semantic-version==2.8.5
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tqdm==4.62.3
typed-ast==1.5.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
varint==1.0.2
vvm==0.1.0
vyper==0.3.1
wcwidth==0.2.5
web3==5.25.0
websockets==9.1
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0
